Build fixes.
```
ResourceManager.cpp:86:21: error: comparison of integers of different signs: 'int' and 'size_type' (aka 'unsigned long') [-Werror,-Wsign-compare]
                for (int i = 0; i < vertexData.size(); ++i) {
                                ~ ^ ~~~~~~~~~~~~~~~~~
ResourceManager.cpp:123:20: error: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
        for (int t = 0; t < triangleCount; ++t) {
                        ~ ^ ~~~~~~~~~~~~~
/Users/mh/privat/motorcycle-game/LearnWebGPU-Code/Application.cpp:334:32: error: cannot take the address of an rvalue of type 'WGPUBindGroupLayout' (aka 'WGPUBindGroupLayoutImpl *')
        layoutDesc.bindGroupLayouts = &(WGPUBindGroupLayout)bindGroupLayout;
                                      ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/Users/mh/privat/motorcycle-game/LearnWebGPU-Code/Application.cpp:358:18: error: use of overloaded operator '=' is ambiguous (with operand types 'wgpu::SwapChainDescriptor' and 'void')
        m_swapChainDesc = {};
```

+++ ResourceManager.cpp
@@ -83,7 +83,7 @@ bool ResourceManager::loadGeometryFromObj(const path& path, std::vector<VertexAt
 		size_t offset = vertexData.size();
 		vertexData.resize(offset + shape.mesh.indices.size());
 
-		for (int i = 0; i < vertexData.size(); ++i) {
+		for (size_t i = 0; i < vertexData.size(); ++i) {
 			const tinyobj::index_t& idx = shape.mesh.indices[i];
 
 			vertexData[offset + i].position = {
@@ -120,7 +120,7 @@ bool ResourceManager::loadGeometryFromObj(const path& path, std::vector<VertexAt
 void ResourceManager::computeTextureFrameAttributes(std::vector<VertexAttributes>& vertexData) {
 	size_t triangleCount = vertexData.size() / 3;
 	// We compute the local texture frame per triangle
-	for (int t = 0; t < triangleCount; ++t) {
+	for (size_t t = 0; t < triangleCount; ++t) {
 		VertexAttributes* v = &vertexData[3 * t];
 
 		// Formulas from http://www.opengl-tutorial.org/intermediate-tutorials/tutorial-13-normal-mapping/
